+++ packages/expo-file-system/src/ExponentFileSystem.ts
@@ -1,7 +1,7 @@
 import { NativeModulesProxy } from 'expo-modules-core';
 
 import ExponentFileSystemShim from './ExponentFileSystemShim';
-import { ExponentFileSystemModule } from './FileSystem.types';
+import { ExponentFileSystemModule } from './types';
 
 let platformModule;