+++ packages/expo-file-system/CHANGELOG.md
@@ -10,6 +10,9 @@
 
 ### 💡 Others
 
+- Extract nested object definitions to the separate types, which adds: `DeletingOptions`, `InfoOptions`, `RelocatingOptions` and `MakeDirectoryOptions` types. ([#20103](https://github.com/expo/expo/pull/20103) by [@Simek](https://github.com/Simek))
+- Simplify the way in which types are exported from the package. ([#20103](https://github.com/expo/expo/pull/20103) by [@Simek](https://github.com/Simek))
+
 ## 15.1.1 — 2022-10-28
 
 _This version does not introduce any user-facing changes._